Validator now ignores in-line DTDs
Sometimes XML files contain DOCTYPE DTD references at the top.
This commit makes the validator ignore these references. This is useful because:
a) We can't guarantee those DTDs will be available on the internet.
b) We can't trust those DTDs to be up to date anyway.

This way, we are only testing the XSD that you choose to use to validate.